Separated main, executable now links to one library.
