+++ README.md
@@ -92,15 +92,13 @@ cmake -DCOMPILER_32=ON ..
 this option on 64-bit architectures will not work since it will not be able to
 store full 64-bit pointers.
 
-#### Build DLL
+#### Shared Library
 
-On Linux, both the static library (.a) and shared library (.so) are generated.
-However, on Windows, only the static library (.lib) is generated by default.
-If you want to create a DLL instead of a static library on Windows, add this
-argument:
+By default, a static library for the compiler is generated (.lib/.a). If you
+want to generate a shared library instead (.dll/.so), add this argument:
 
 ```bash
-cmake -DCOMPILER_DLL=ON ..
+cmake -DCOMPILER_SHARED=ON ..
 ```
 
 #### Enable C API Tests
