+++ tests/c/CMakeLists.txt
@@ -22,11 +22,11 @@
 # of executables made here.
 function(link_with_decision testTarget)
     if(MSVC)
-        if(COMPILER_DLL)
+        if(COMPILER_SHARED)
             target_link_libraries(${testTarget} decisionLibShared)
-        else(COMPILER_DLL)
+        else(COMPILER_SHARED)
             target_link_libraries(${testTarget} decisionLibStatic)
-        endif(COMPILER_DLL)
+        endif(COMPILER_SHARED)
     else(MSVC)
         target_link_libraries(${testTarget} decisionLibShared)
     endif(MSVC)
